Login to LaminDB

Login as a LaminDB user

Usage

lamin_login(user = NULL, api_key = NULL)

Arguments

user

Handle for the user to login as

api_key

API key for a user

Details

Setting user will run lamin login <user>. Setting api_key will set the LAMIN_API_KEY environment variable temporarily with withr::with_envvar() and run lamin login. If neither user or api_key are set the user handle will be retrieved from the user settings file. If that is not possible lamin login will only be run if LAMIN_API_KEY is set.